change dir and add rounding
