--- flowers102/mobilenet/README.md
@@ -1,12 +1,0 @@
-## Mobilenet trained on flower102 
-The trained model can be download from here ([trained model](https://pan.baidu.com/s/1geHkrw3)). The accuracy is 97.16%.
-
-
-### Reproduce 
-1. download the pretraind model.
-
-	the `imagenet_pretrained_mobilenet.tar.gz`( [Download from BaiduCloud](https://pan.baidu.com/s/1slK3wlB)) is the pretrained model on imagenet, top-1 accuracy : 70.51%
- top-5 accuracy 89.85%
-
-2. train the model   
-`python train.py`
