+++ models/resnet18/README.md
@@ -1,0 +1,12 @@
+## Resnet18 trained on flower102 
+The trained model can be download from here ([trained model](https://pan.baidu.com/s/1pLlwg9X)). The accuracy is 97.353%.
+
+
+### Reproduce 
+1. download the pretraind model.
+
+	the `imagenet_pretrained_resnet18.tar.gz`( [Download from BaiduCloud](https://pan.baidu.com/s/1gf7rd0v)) is the pretrained model on imagenet, top-1 accuracy : 70.91%
+ top-5 accuracy 89.6%
+
+2. train the model   
+`python train.py`
